Calender context moved into redux.
